+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0]
+
+### Added
+* Added a new parameter to include PS processing `--process` that can be `ps` or `sbas`.
+* Added `--tbaseline` and `--pbaseline` parameters to specify temporal and perpendicular baselines respectively.
+
 ## [0.9.9]
 
 ### Changed